browser-use-tests/pyproject.toml
[project]
name = "browser-use-tests"
version = "0.1.0"
description = "Browser-Use agent for testing vulnerable applications"
requires-python = ">=3.11"
dependencies = [
    "browser-use==0.10.1",
    "playwright>=1.56.0",
    "python-dotenv>=1.0.0",
]

[tool.uv]
package = false
